Remove sorting
Sorting is too difficult to undo consistently with multiple of the same
character in the key. A possible replacement is a conveyer belt
algorithm in which each row is shifted by an amount determined from its
key letter. Any that would be past the end of the row loop back to the
beginning, like a conveyor belt.